+++ stories/Hooks/index.js
@@ -3,5 +3,6 @@ import UseFormState from './useFormState';
 import UseFieldApi from './useFieldApi';
 import UseFieldState from './useFieldState';
 import UseArrayField from './useArrayField';
+import UseForm from './useForm';
 
-export { UseFormApi, UseFormState, UseFieldApi, UseFieldState, UseArrayField };
+export { UseFormApi, UseFormState, UseFieldApi, UseFieldState, UseArrayField, UseForm };